Make this an Express app
We'll need to have dynamic functionality, so make this a dynamic
app using the Express framework.

Implement a basic API for creating and returning schools.

Render home page using Handlebars templates.

Update front-end build scripts to write to `public` directory.

Use the API in an npm script to initialize the schools from CSV
instead of generating a GeoJSON file.