Add program through REST API
Update the REST API endpoint for /schools/:rcdts/programs to handle the
`age_group` and `program_type` parameters.

Add an API wrapper module to abstract away the request/response
handling.  Use a polyfill for the forthcoming fetch API instead of using
xhr-promise.

Add a `createProgram` action creator that sends a request to the REST
API and then calls the `receiveProgram` action creator.

Add a `receiveProgram` action in a separate actions module,
`LearningCollaborativeServerActions`.  This actions module can be used
to create actions from server responses.

Update SchoolStore to emit an event when a program is received via the
aforementioned action.

Also add a list of school program types to SchoolStore and pass this
thorugh the component hierarchy, as props, to `CreateProgram`.

Add a `CreateProgram` component to render the form to add a program to a
school.  This is rendered as a child of `SchoolDetail`.

Addresses https://github.com/ghing/learning-collaborative-maps/issues/1