Add school data to repo
This isn't great, but the geocoding is an expensive enough
process and the school data will change infrequently enough
that it makese sense to add this to the repo, if only to make
it easier for me to initialize different dev environments.

Update the npm scripts related to the data pipeline so they write
and load the schools from the CSV in the repo.