Start implementing program editing
Add a component for editing a program, `EditProgram`.  It's pretty
similar to `CreateProgram`, so this should probably be refactored later.

Create a `utils` module and factor out agency ID parsing into a function
there.

Update the `SchoolProgram` component to show an `edit` link on hover
that will direct `SchoolDetail` to show the editing form.

Update `SchoolDetail` to toggle hiding the program list and showing the
edit program form.

Addresses #28