Implement updating the program note
Wire up route and database API function for updating a program
note.

Update the `updateSchoolProgram` database API function so that it returns the full program object, including the notes, that was updated, as the object included in request might not contain all the fields.

Include a method in the `receiveProgramNote` server action so note creation and update can be handled differently.

Add a method to `LearningCollaborativeMap` to handle receiving a note.

Make `SchoolStore` emit an event when the `RECEIVE_PROGRAM_NOTE` action happens.

Addresses #28