Use actions to change program mode
The application currently has 3 modes:

* Viewing a school detail
* Adding a school program
* Editing a school program

Try to make this state live at the top level (in SchoolStore)
and passed down as props instead of living in a downstream view.

Change the state using actions.

This seems like a more Fluxy way to do it, and is ultimately
easier for me to reason about.

All of this was changes that I noticed should be made on the way
to reseting back to showing a school detail when a new school is
selected.

Addresses #41