Make it easier to click school markers
Sometimes clicking on a marker for a school doesn't show the school
details in the sidebar.  I've speculated that this is because the
marker is too small, making it easier to click on the edge or
outside of the marker even when aiming for the marker and because
of the placement of the popup, making it possible to click on the
popup element when you think you're clicking on the marker.

To address this, make the markers larger (9px radius instead of 6px)
and make the offset of the tooltip from the marker greater.

Addresses #36
https://github.com/ghing/learning-collaborative-maps/issues/36