Don't initialize Typeahead plugin multiple times
The [typeahead.js](https://twitter.github.io/typeahead.js/) jQuery
plugin was getting initialized multiple times on the same element.  This
was causing multiple dropdown elements to appear on subsequent searches
and for the search input text to overlap and appear garbled.

By checking for a property set with jQuery's
[.data()](https://api.jquery.com/data/), I can see if the typeahead
plugin has been initialized yet on a particular element and only
initialize the plugin if it hasn't been initialized.

Addresses #50
https://github.com/ghing/learning-collaborative-maps/issues/50