Better mouse interactions for search suggestions
Make it easier to see which item is being selected when using the
mouse with the suggestions in the school search input.

Do this by highlighting the search suggestions with a background
color when it is hovered over.

Also change the cursor style to be the default (an arrow) instead
of the text selection icon.

Addresses #58
https://github.com/ghing/learning-collaborative-maps/issues/58