Routes for schools, adding and editing programs
Implement client and server-side routes for viewing a school's programs,
adding a program and editing a program using React Router.

This give us clean URLs for specific app states, which is just a nice,
traditional way for a web app to work.  It also lets user bookmark a
specific school, which could save some for users who only need to update
a couple of schools.

Since we now have routes, we can use navigation, either through the router
object or through the history object, to indicate how we want to change
state.  So we can eliminate actions related to switching between views
showing the school detail, the form to add a program and the form to edit
a program.  As a result, the notion of the application "mode" and the
event emitters and listeners on `SchoolStore` related to this go away.

The router config defines the nesting of components based on the route that
you visit.  So, in many of the components such as `LearningCollaborativeMap`
and `SchoolDetail` we pass through `this.props.children` instead of
instantiating the child components in the render method.  However we want
to pass on additional props that are known to the parent component, but not
the router config.  To do this, we use `React.Children.map` and
`React.cloneElement`.

Update `ProgramForm` to just accept the program to be edited as a prop
rather than separate initial values for each of the fields.  The state
values for each form element is just derived from the program. This avoids
branching when passing props to this component from its parent.

Since we now have subpaths, update the server-side templates to link to
the CSS and JavaScript bundles using a leading '/'.  That is, they're
absolute paths.

Use ES6 class definitions for the React components we touch in this commit
rather than `React.createClass`.  This just seems like the preferred
approach that I'm seeing in other React examples.

Addresses #65
https://github.com/ghing/learning-collaborative-maps/issues/65