Implement a login flow
It's not create to have an app where we manage data hanging out with no
authentication.  However, people hate having to remember yet another
password.  Use a token-based authentication scheme like Slack does.
Use the [Passwordless](https://passwordless.net/) library to save us a
lot of hard work.

This adds a couple of new dependencies.  We have to use
[express-session](https://github.com/expressjs/session) for session
management, and we use MongoDB to store the session tokens so they
can persist across app restarts.

Passwordless authentication also requires sending the token through
e-mail, so we use Sendgrid and an API library to help with that.

Implement the UI components to allow users to login using a passwordless
authentication scheme.

Add a new store, `UserStore`, and some action creators to implement this
flow.  Also implement an Express middleware to provide an API endpoint
to request the token.  This still needs to be implemented.

Addresses #60
https://github.com/ghing/learning-collaborative-maps/issues/60