Add logout route, redirect to login
Redirect to the login route when a user is not authenticated instea of
just showing a mysterious "Unauthroized" message.

Also add a `/logout` route.  In practice, I can't imagine users using
this very much, but it's useful for testing.  I guess it's also useful
if multiple users are using the same device.

Addresses #60
https://github.com/ghing/learning-collaborative-maps/issues/60