Interface for adding agencies
Implement React components and routes for an interface for listing
and adding new agencies.

Fix tests so they run. This required updating Jest to the latest version
which no longer automocks by default and changing the test filenames and
requires to match the current module names.

Add documentation on running tests so that I can remember how to run
tests since I touch this project so sporadically.

Document the data flow through the app in the README.  Things can get
pretty confusing quickly in a Flux app, so it helps to have an
understanding of how data flows through the app.

Addresses #59