Lift up state when editing agencies
Keep state for the actively edited or created agency in AgencyAdmin
rather than AgencyAdminForm.

This allows the agency's fields to be visible whether clicking on a link
from the list view or visiting `/admin/agencies/:slug` directly.

For more on lifting state up, see this page in the React documentation:
https://facebook.github.io/react/docs/lifting-state-up.html.

Addresses #71