Modernize some dependencies
Use the new, official sendgrid library.

Use a version of the MongoDB token store that uses node-bcrypt, because
the version of bcrypt required by the package can't be built with newer
versions of Node.js. I still had to fork this package, becaue the
version of the mongodb package didn't work with the version of Mongo on
the server of the hosted Atlas service.